# [rgw]: Fast path for resharding