@cbodley 2024-04-11
thanks for the contribution, i'm interested to learn more about this design! i see that the pr targets the quincy branch though - it will need to be rebased on the latest main branch. this is why i added the 'do-not-merge' label

there's another pr https://github.com/ceph/ceph/pull/56597 that i think is trying to accomplish the same thing. hopefully we can take the best ideas from both approaches (cc @liangmingyuanneo)