@cbodley 2024-04-11
thanks @YaZasnyal, parallelism is a big improvement here :rocket:

the work in https://github.com/ceph/ceph/pull/56597 adds a separate stage to the reshard algorithm, and only blocks writes during that second stage. i believe that both stages of that approach would benefit from this kind of shard parallelism

> This process can be done in parallel for each shard if the new number of shards is multiple of the old one. This way modulo operation is going to fall into exactly n new shards where n is growth factor.

can you please expand on this part? the resharding logic should be independent across shards, so i'd expect this strategy to work regardless of the number of source/target shards