@YaZasnyal 2024-04-12
> i see that the pr targets the quincy branch though - it will need to be rebased on the latest main branch. this is why i added the 'do-not-merge' label

Yes we still use quincy and I decided to share my thoughts on this subject and get some feedback. I think I can open a new one for the main branch after the initial design is finished.

> can you please expand on this part? the resharding logic should be independent across shards, so i'd expect this strategy to work regardless of the number of source/target shards

I assumed that one rados object should not be accessed from multiple threads but now I am not sure about that. If this is not the case we don't need this extra path and can do it in parallel for any number of shards.