@cbodley 2024-04-12
> I assumed that one rados object should not be accessed from multiple threads but now I am not sure about that. If this is not the case we don't need this extra path and can do it in parallel for any number of shards.

it's fine for multiple threads (or multiple librados clients) to send writes to the same object. the osd makes sure that each write is applied atomically